refactor(ParticipantHand) : 참가자의 핸드정보 클래스 분리
